Fix moving memory store code into correct mod
